Update the README.md file to be consistent with the work i did in the Bikeshare project
I have updated and added the README.md file for this project of version control to be consistent with
the work i did in the bikeshare project. I have added all the required and necessary description below
the headings and explained the bikeshare project, data sets, question that have been asnwered by the
project and the reseources reffered to complete the project.